backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Flask>=3.0.2,<4.0
elasticsearch>=7.17.0,<8.0
SQLAlchemy>=2.0.27,<3.0
gunicorn==20.1.0
requests>=2.25.1
Flask-SQLAlchemy
psycopg2-binary==2.9.3
python-dotenv
flask-cors